reset cfg if empty, define destination=
this change implement 2 updates.

force configure if cfg is empty. got twice for some reason a cfg file empty, and next run was failing with a hard to find issue.

if destination= is defined, use pushd and popd to change to that directory. This allow run this script from crontab or from other scripts.